requirements.txt
aiodns==3.0.0
aiosignal==1.2.0
aniso8601==7.0.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
Brotli==1.0.9
cached-property==1.5.2
cchardet==2.1.7
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
Django==4.0.6
django-filter==22.1
django-graphql-auth==0.3.15
django-graphql-jwt==0.3.0
djangorestframework==3.13.1
frozenlist==1.3.0
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
idna==3.3
iso8601==1.0.2
isodate==0.6.1
lxml==4.9.1
multidict==6.0.2
Pillow==9.2.0
platformdirs==2.5.2
promise==2.3
pycares==4.2.1
pycparser==2.21
PyJWT==2.4.0
pytz==2022.1
requests==2.28.1
requests-file==1.5.1
requests-toolbelt==0.9.1
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
urllib3==1.26.10
yarl==1.7.2
zeep==4.1.0
